add and change
add vent object. allow view tell each view when rendering collection to
trigger an event when finishing rendering. count the number of event and
when we get to zero - order the columns
change: margin size is auto calc also the number of columns.
todo: allow few types of action 1. auto calc of columns and margin, 2.
fixed margin. 3. fixed number of columns and so on .maybe event big item
on two columns (or more) together)